Expiry time is reduces

+++ src/services/campaign-service.ts
@@ -79,7 +79,7 @@ export const completeCampaignOperation = async (card: campaign_twittercard) => {
       await updateAllEngagementsForCard(card),
     ]);
 
-    const campaignExpiry = moment().add(parseInt(process.env.REWARD_CALIM_HOUR!), "hours").toISOString();
+    const campaignExpiry = moment().add(parseFloat(process.env.REWARD_CALIM_HOUR!), "hours").toISOString();
     //log campaign expiry
     logger.info(`Campaign expired at ${campaignExpiry}`);
